Rename module and remove unused import.
