+++ python/__init__.py
@@ -8,4 +8,4 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
-from . import tk_multi_launchapp  # , app_launch_overlay
+from . import tk_multi_launchapp
