+++ python/launch_indicator_dialog/resources/build_resources.sh
