+++ python/launch_indicator_dialog/resources/splash_new.ui
