+++ python/tk_multi_launchapp/__init__.py
@@ -10,6 +10,5 @@
 
 from .single_config_launcher import SingleConfigLauncher
 from .software_entity_launcher import SoftwareEntityLauncher
-from ..app_launch_overlay import qtwidgets
+from ..launch_indicator_dialog import qtwidgets
 
-# from .qtwidgets import ShotgunSpinningWidget, overlay_widget
